+++ controllers/codeforces_controller.js
@@ -6,6 +6,14 @@ const calculator = require('../rankCalculator/rankCalculator');
 
 module.exports.enterData = async function(req, res){
     try{
+
+        if(req.body.my != req.user._id)
+        {
+            return res.json(500, {
+                message: 'Invalid Access'
+            });
+        }
+
         let user = await User.findById(req.user._id);
 
         let codeforces = await CodeForces.findOne({user: req.user._id});
@@ -96,6 +104,14 @@ module.exports.enterData = async function(req, res){
 module.exports.removeData = async function(req, res){
 
     try{
+
+        if(req.query.Id != req.user._id)
+        {
+            return res.json(500, {
+                message: 'Invalid Access'
+            });
+        }
+
         let codeforces = await CodeForces.findOneAndDelete({user: req.user._id});
 
         let rank;
